move share import implementation to pkg
